+++ src/inputscope/conf.py
@@ -23,6 +23,7 @@ the declared ones in source code. File is deleted if all values are at default.
 @modified    06.07.2022
 ------------------------------------------------------------------------------
 """
+import ast
 try: import ConfigParser as configparser # Py2
 except ImportError: import configparser  # Py3
 import datetime
@@ -35,7 +36,7 @@ import sys
 
 """Program title, version number and version date."""
 Title = "InputScope"
-Version = "1.6.dev7"
+Version = "1.6.dev8"
 VersionDate = "06.07.2022"
 
 """TCP port of the web user interface."""
@@ -322,15 +323,24 @@ DbUpdateStatements = [
         "ALTER TABLE scrolls ADD COLUMN dx INTEGER DEFAULT 0"]],
 ]
 
+"""List of attribute names that are always saved to ConfigFile."""
+FileDirectives = ["CustomKeys", "DefaultScreenSize", "EventsWriteInterval", "MaxEventsForStats",
+    "MaxEventsForReplay", "KeyboardEnabled", "KeyboardKeysEnabled", "KeyboardCombosEnabled",
+    "MouseEnabled", "MouseMovesEnabled", "MouseClicksEnabled", "MouseScrollsEnabled",
+    "MouseHeatmapSize", "MouseMoveJoinInterval", "MouseMoveJoinRadius", "MouseScrollJoinInterval",
+    "PixelLength", "ScreenSizeInterval", "WebPort",
+]
 
 try: text_types = (str, unicode)       # Py2
 except Exception: text_types = (str, ) # Py3
 
-def init(filename=ConfigPath):
-    """Loads INI configuration into this module's attributes."""
+def init(filename=ConfigPath, create=True):
+    """Loads INI configuration into this module's attributes; creates file if not present."""
     section, parts = "DEFAULT", filename.rsplit(":", 1)
     if len(parts) > 1 and os.path.isfile(parts[0]): filename, section = parts
-    if not os.path.isfile(filename): return
+    if not os.path.isfile(filename):
+        if create: save()
+        return
 
     vardict, parser = globals(), configparser.RawConfigParser()
     parser.optionxform = str # Force case-sensitivity on names
@@ -338,7 +348,7 @@ def init(filename=ConfigPath):
         def parse_value(raw):
             try: return json.loads(raw) # Try to interpret as JSON
             except ValueError:
-                try: return ast.literal(raw) # JSON failed, fall back to eval
+                try: return ast.literal_eval(raw) # JSON failed, fall back to eval
                 except (SyntaxError, ValueError): raw # Fall back to raw
         with open(filename, "r") as f:
             txt = f.read()
@@ -360,20 +370,14 @@ def save(filename=ConfigPath):
     try:
         save_types = text_types + (int, float, tuple, list, dict, type(None))
         for k, v in sorted(globals().items()):
-            if not isinstance(v, save_types) or k.startswith("_") \
-            or k not in default_values or default_values[k] == v \
-            or isinstance(default_values[k], tuple) and isinstance(v, list) \
-            and default_values[k] == tuple(v): continue # for k, v
+            if not isinstance(v, save_types) or k.startswith("_") or k not in default_values \
+            or k not in FileDirectives and default_values[k] == v: continue # for k, v
             try: parser.set("DEFAULT", k, json.dumps(v))
             except Exception: pass
-        if parser.defaults():
-            with open(filename, "w") as f:
-                f.write("# %s %s configuration written on %s.\n" % (Title, Version,
-                        datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")))
-                parser.write(f)
-        else: # Nothing to write: delete configuration file
-            try: os.unlink(filename)
-            except Exception: pass
+        with open(filename, "w") as f:
+            f.write("# %s %s configuration written on %s.\n" % (Title, Version,
+                    datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")))
+            parser.write(f)
     except Exception:
         logging.warn("Error writing config to %s.", filename, exc_info=True)
 
@@ -381,6 +385,11 @@ def save(filename=ConfigPath):
 def validate():
     """Validates configuration values, discarding invalids."""
     global CustomKeys
+    default_values = defaults()
+    for k, v in sorted(globals().items()):
+        v0 = default_values.get(k)
+        if isinstance(v, list) and isinstance(v0, tuple):
+            globals()[k] = tuple(v)
     try:
         keys, _ = CustomKeys.copy(), CustomKeys.clear()
         for k, v in keys.items():